# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Staging Superapp API",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server (Docker)"
        }
    ],
    "paths": {
...
        "/api/v1/inventory/stocks": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get list of inventory stocks",
                "operationId": "2e65f0da91b517ed7db2dec233821cac",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InventoryStock"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar stok berhasil diambil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/stocks/{itemId}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get stock details by item ID",
                "operationId": "06cc48b8e755880fda00b0d637b45ac9",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "ID of the item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InventoryStock"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail stok per item berhasil diambil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/movements": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get stock movements history",
                "operationId": "fee3b12165a8de5e901d40494c780631",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Riwayat pergerakan stok berhasil diambil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InventoryStock": {
                "title": "Inventory Stock Schema",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": 1
                    },
                    "item_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "location_id": {
                        "type": "string",
                        "example": 1
                    },
                    "bin_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "nullable": true
                    },
                    "batch_no": {
                        "type": "string",
                        "example": "BATCH-001",
                        "nullable": true
                    },
                    "serial_no": {
                        "type": "string",
                        "example": "SN-001",
                        "nullable": true
                    },
                    "expired_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31",
                        "nullable": true
                    },
                    "qty": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-04T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}